# delilah_Resto_Project/project/api.yaml
swagger: "2.0"
info:
  description: "This is the Delilah Restó Restaurant API. Where you can make your order online, and receive it in your home, easy and quick."
  version: "1.0.0"
  title: "Delilah Restó"
  termsOfService: "http://delilah/terms/"
  contact:
    email: "delilahresto@gmail.com.co"
  license:
    name: "Apache 2.0"
    url: "http://www.delilahresto.com.co"
host: "delilah.swagger.io"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /orderDetails:
    post:
      tags:
      - "orderDetails"
      summary: "Create the order detail"
      description: ""
      operationId: "logoutUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "order"
        description: "JSON object that needs to create a new order detail"
        required: true
        schema:
          type: "object"
          properties:
            order_id:
              type: "integer"
            products:
              type: "object"
              properties:
                product_id:
                  type: "integer"
                quantity:
                  type: "integer"
      - in: "header"
        type: "string"
        name: "token"
        description: "Token to authenticate."
        required: true            
      responses:
        "200":
          description: "successful operation"
    put:
      tags:
      - "orderDetails"
      summary: "Get user by user name"
      description: ""
      operationId: "getUserByName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: ""
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
    delete:
      tags:
      - "orderDetails"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      user_name:
        type: "string"
      full_name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "integer"
        format: "int64"
      address:
        type: "string"
      password:
        type: "string"
      admin:
        type: "boolean"
        default: false
        description: "User Status"
    xml:
      name: "User"
...